rmg: the dwarven underground closes - 505,183 draws and every object
Both of the two open draws came off the forced-seed trace of GCK-011,
and neither was where the log said it was.

The FillTerrain pre-step: GenerateMap gates it on the dwarven coin
(cmp byte ptr [ecx+8Ch],0 at 0xEABA09) and 0xED17F0 runs 0xEB2A20 on
the level's SECOND floor, present only when the floor vector holds two
(end - begin >= 0x240, entries of 0x120). The draw is the literal 8 of
0xEB2A25, smeared as 8 + below(8) over a (w/3+1)x(h/3+1) coarse grid -
and the traced map has eight zones, so a limit of 8 in the log read
equally well as below(zones). dwarvenCoarse existed and had never been
called; the chain calls it now, between calcBorderTiles and placeTowns.

The treasure blocks were not a treasure-block rule. The engine grew the
block at seed (3,55) with TWO points and the port with three, and with
three the value per point fell under 2, so the port laid a chest where
the engine drew a resource kind - the missing below(7) was the shape of
the difference, not its cause. The extra point passed the growth's "two
footprints around it" on tiles at x = 1, which carry no object at all in
the game's map. Those are the dwarven rock mask's tiles: 0xEC7225 sets
0x40 AND 0x400 on them, and the massif carve of the next dwarven zone
converts a tile reading EXACTLY 0x40 into a footprint (cmp dword ptr
[eax+ecx*4],40h, 0xED162A). The port's occupancy was a byte, the 0x400
fell off it, and the carve claimed tiles the engine leaves alone. The
grid is an Int32Array now - which is what the engine has, or dword ptr
[eax+ebx*4],400h - the mask sets both bits, the marked list is taken by
0x400, and the deep columns are 2 | 0x400.

GCK-011 now matches the game draw for draw, all 505,183, and its
map.xdb - 9,943 objects - is byte-identical. Eighteen of its twenty
entries are; the two that are not, UndergroundTerrain.bin and
minimap_floor_02.dds, are the same two on GCK-004 and GCK-015 and on no
other map, and all three of those are the dwarven ones. Over the twelve
game maps: seven byte-identical in every entry, those three at 18 of 20,
GCK-007 at 19 (the Fairie Tree icon), GCK-001 unchanged at 14 of 17,
measured again with the change stashed. test-rmg-underground holds the
two occupancy values apart without any game data, and reddens when
either half is undone.

+++ src/rmg/cartographer.ts
@@ -35,7 +35,7 @@ export interface CartographerInput {
   grid: Int32Array[];
   border: Int32Array[];
   /** MUTATED: the stamp marks 2 and 4. */
-  occupancy: Uint8Array;
+  occupancy: Int32Array;
   /** MUTATED: the stamp's 4s join the zone's room points. */
   points: Tile[];
   /** MUTATED when carried: stamped-blocked tiles join the zone's `+0x5C` ledger. */
