+++ src/rmg/mines.ts
@@ -132,7 +132,7 @@ export interface MineStepInput {
   grid: Int32Array[];
   border: Int32Array[];
   /** MUTATED: the step marks 2 and 4 the way the engine does. */
-  occupancy: Uint8Array;
+  occupancy: Int32Array;
   /** MUTATED: the zone's stamped points — what the room is measured from. */
   points: Tile[];
   /** MUTATED when carried: stamped-blocked tiles join the zone's `+0x5C` ledger. */
@@ -276,7 +276,7 @@ export interface AbandonedMinesInput {
   grid: Int32Array[];
   border: Int32Array[];
   /** MUTATED: the stamp marks 2 and 4. */
-  occupancy: Uint8Array;
+  occupancy: Int32Array;
   /** MUTATED: the stamp's 4s join the zone's room points. */
   points: Tile[];
   /** MUTATED when carried: stamped-blocked tiles join the zone's `+0x5C` ledger. */
