+++ src/rmg/prisons.ts
@@ -40,7 +40,7 @@ export interface PrisonsInput {
   grid: Int32Array[];
   border: Int32Array[];
   /** MUTATED: the stamp marks 2 and 4. */
-  occupancy: Uint8Array;
+  occupancy: Int32Array;
   /** MUTATED: the stamp's 4s join the zone's room points. */
   points: Tile[];
   /** MUTATED when carried: stamped-blocked tiles join the zone's `+0x5C` ledger. */
