+++ src/rmg/upgrade-buildings.ts
@@ -73,7 +73,7 @@ export interface SeatedGuard {
 export function seatGuard(
   input: {
     size: number;
-    occupancy: Uint8Array;
+    occupancy: Int32Array;
     at: Tile;
     q: number;
     foot: Footprint;
@@ -116,7 +116,7 @@ export interface UpgradeBuildingsInput {
   grid: Int32Array[];
   border: Int32Array[];
   /** MUTATED: the stamp marks 2 and 4; a seated guard marks 4. */
-  occupancy: Uint8Array;
+  occupancy: Int32Array;
   /** MUTATED: the stamp's 4s join the zone's room points. */
   points: Tile[];
   /** MUTATED when carried: stamped-blocked tiles join the zone's `+0x5C` ledger. */
